+++ articles/extensions/thumbnail.py
@@ -1,11 +1,21 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
+from feincms import extensions
 
-def register(cls, admin_cls):
-    cls.add_to_class('thumbnail', models.ImageField(_('thumbnail'), max_length=250, upload_to="images/articles/thumbnails", null=True, blank=True))
-    if admin_cls:
-        admin_cls.add_extension_options(_('Thumbnail'), {
+
+class Extension(extensions.Extension):
+    def handle_model(self):
+        self.model.add_to_class(
+            'thumbnail',
+            models.ImageField(
+                _('thumbnail'),
+                max_length=250,
+                upload_to="images/articles/thumbnails",
+                null=True,
+                blank=True)
+        )
+
+    def handle_modeladmin(self, modeladmin):
+        modeladmin.add_extension_options(_('Thumbnail'), {
             'fields': ('thumbnail',),
         })
-
-